refactor: 프로토콜 및 node 별도 파일로 분리
