+++ Calculator/Calculator/Model/CalculateItem.swift
@@ -1,0 +1,10 @@
+//
+//  CalculateItem.swift
+//  Calculator
+//
+//  Created by JeongTaek Han on 2021/11/11.
+//
+
+import Foundation
+
+protocol CalculateItem { }
